build both normal configuration and windows service configuration to their own seperate executables for easy testing
