+++ Cargo.toml
@@ -2,6 +2,7 @@
 name = "actix-svelte"
 version = "0.1.0"
 edition = "2021"
+author = "Sergio Falcon <sergiojfalcon@gmail.com>"
 
 [dependencies]
 actix-files = "0.6.6"
@@ -12,5 +13,17 @@ mime_guess = "2.0.5"
 rust-embed = { version = "8.5.0", features = ["actix", "actix-web"] }
 serde = { version = "1.0.217", features = ["derive"] }
 serde_json = "1.0.135"
-tokio = { version = "1.42.0", features = ["full"] }
+tokio = { version = "1.43.0", features = ["full"] }
 windows-service = "0.7.0"
+
+[features]
+default = [] # No features enabled by default
+windows_service = [] # Enables service mode
+
+[[bin]]
+name = "actix-svelte"
+path = "src/main.rs"
+
+[[bin]]
+name = "actix-svelte-ws"
+path = "src/ws-main.rs"
